requirements.txt
streamlit==1.32.2
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
requests==2.31.0
python-dotenv==1.0.1 